+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 2.8.0 (Future)
 - Allow more than 1 date filter on a listing
 - Add toggle to date filter for single date
+- Alphabetically sort menu items
 
 ## Version 2.7.1 (2017-09-25)
 - Fix select2 class on BelongsToField