+++ changelog.md
@@ -4,6 +4,7 @@
 - Allow more than 1 date filter on a listing
 - Add toggle to date filter for single date
 - Alphabetically sort menu items
+- Add name and search filter to user listing
 
 ## Version 2.7.1 (2017-09-25)
 - Fix select2 class on BelongsToField