+++ changelog.md
@@ -5,6 +5,7 @@
 - Add toggle to date filter for single date
 - Alphabetically sort menu items
 - Add name and search filter to user listing
+- Fix per page select on media library listing
 
 ## Version 2.7.1 (2017-09-25)
 - Fix select2 class on BelongsToField